renamed to org.eclipse.smarthome.automation.sample.handler
