+++ bundles/automation/org.eclipse.smarthome.automation.sample.handler/.classpath
