+++ bundles/automation/org.eclipse.smarthome.automation.sample.handler/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.eclipse.smarthome.automation.sample.handler.factories</name>
+	<name>org.eclipse.smarthome.automation.sample.handler</name>
 	<comment></comment>
 	<projects>
 	</projects>
