+++ bundles/automation/org.eclipse.smarthome.automation.sample.handler/ESH-INF/automation/rules/DemoScriptRule.json
