+++ bundles/automation/org.eclipse.smarthome.automation.sample.handler/ESH-INF/automation/templates/SampleTemplateDefinitions.json
