+++ bundles/automation/org.eclipse.smarthome.automation.sample.handler/ESH-INF/i18n/resourceName_bg.properties
