+++ bundles/automation/org.eclipse.smarthome.automation.sample.handler/META-INF/MANIFEST.MF
@@ -1,8 +1,8 @@
 Bundle-ManifestVersion: 2
-Bundle-SymbolicName: org.eclipse.smarthome.automation.sample.handler.factories
+Bundle-SymbolicName: org.eclipse.smarthome.automation.sample.handler
 Bundle-Version: 0.8.0.qualifier
 Bundle-Vendor: Eclipse.org/SmartHome
-Bundle-Name: Automation sample.handler.factories
+Bundle-Name: Automation sample.handler
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Import-Package: 
  org.eclipse.smarthome.automation,
@@ -17,4 +17,4 @@ Import-Package:
  org.osgi.framework,
  org.osgi.service.event,
  org.osgi.util.tracker;version="1.4.0"
-Bundle-Activator: org.eclipse.smarthome.automation.internal.sample.handler.factories.Activator
+Bundle-Activator: org.eclipse.smarthome.automation.internal.sample.handler.Activator
