+++ bundles/automation/org.eclipse.smarthome.automation.sample.handler/SmartHome Runtime Sample Rules.launch
