+++ bundles/automation/org.eclipse.smarthome.automation.sample.handler/build.properties
