+++ bundles/automation/org.eclipse.smarthome.automation.sample.handler/src/main/java/org/eclipse/smarthome/automation/internal/sample/handler/SampleActionHandler.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.automation.internal.sample.handler.factories;
+package org.eclipse.smarthome.automation.internal.sample.handler;
 
 import java.util.Map;
