+++ bundles/automation/pom.xml
@@ -22,7 +22,7 @@
     <module>org.eclipse.smarthome.automation.providers</module>
     <module>org.eclipse.smarthome.automation.core</module>
     <module>org.eclipse.smarthome.automation.core.test</module>
-    <module>org.eclipse.smarthome.automation.sample.handler.factories</module>
+    <module>org.eclipse.smarthome.automation.sample.handler</module>
     <module>org.eclipse.smarthome.automation.module.core</module>
     <module>org.eclipse.smarthome.automation.module.core.test</module>
     <module>org.eclipse.smarthome.automation.module.script</module>
